clean up for getCategories()
